폴더 이름 memory로 바꿈
