+++ memory/app/chapter/errors.py
@@ -1,0 +1,2 @@
+from mysol.common.errors import MysolHTTPException
+
