+++ memory/common/errors.py
