+++ memory/database/common.py
