+++ memory/database/middleware.py
