--- mysol/app/user/hashing.py
@@ -1,13 +1,0 @@
-from passlib.context import CryptContext
-
-# Bcrypt 해싱 설정
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-class Hasher:
-    @staticmethod
-    def verify_password(plain_password: str, hashed_password: str) -> bool:
-        return pwd_context.verify(plain_password, hashed_password)
-
-    @staticmethod
-    def hash_password(password: str) -> str:
-        return pwd_context.hash(password)
